relocated images. new close icon. renamed "testpath" into "path". added close button and related method
