--- Rocket tracking widget - to share/launcher.py
@@ -1,5 +1,0 @@
-import subprocess
-import os
-
-script_path = os.path.join(os.path.dirname(__file__), "Rocket widget.py")
-subprocess.Popen(["start", "pythonw", script_path], shell=True)
